chore(web): move fileUtils.ts to lib/projects/utils.ts
